End update functions

+++ BL/CustomerInParcel.cs
@@ -4,7 +4,7 @@ namespace IBL.BO
 {
     public class CustomerInParcel
     {
-        public ulong id { get; set; }
+        public int id { get; set; }
         public string name { get; set; }
     }
 }
